feat(enums): add an enum to modify the behaviour and its controller

+++ csharp-delegates_events/3-modified_behavior/3-modified_behavior.cs
@@ -94,7 +94,7 @@ public class Player {
         } else if (mod == Modifier.Base) {
             return baseV;
         } else {
-            return (float)(baseV * 1.5);
+            return baseV * 1.5f;
         }
     }
